% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code2.R
\name{ExValue_from_Pointshp}
\alias{ExValue_from_Pointshp}
\title{At the time of use this you have to add a Point shapefile.}
\usage{
ExValue_from_Pointshp(ras_img, point_shp, res_Col_name)
}
\arguments{
\item{ras_img}{Raster image}

\item{point_shp}{Point shapefile with class info}

\item{res_Col_name}{Name of the colomn contain class id's}
}
\description{
At the time of use this you have to add a Point shapefile.
}
\examples{
library(raster)
library(ExtractTrainData)
ras_img<-brick(system.file("extdata","ras.tif",package = "ExtractTrainData"))
point_shp<-shapefile(system.file("extdata","poin.shp",package = "ExtractTrainData"))
res_Col_name<-"Id"
ExValue_from_Pointshp(ras_img,point_shp,res_Col_name)
}
\author{
Subhadip Datta
}
