% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{ExValue_from_Polyshp}
\alias{ExValue_from_Polyshp}
\title{Add a Polygon shapefile and raster image.}
\usage{
ExValue_from_Polyshp(ras_img, poly_shp, res_Col_name)
}
\arguments{
\item{ras_img}{Raster image}

\item{poly_shp}{Polygon shapefile with class info}

\item{res_Col_name}{Name of the colomn contain class id's}
}
\description{
Add a Polygon shapefile and raster image.
}
\examples{
library(raster)
library(ExtractTrainData)
ras_img<-brick(system.file("extdata","ras.tif",package = "ExtractTrainData"))
poly_shp<-shapefile(system.file("extdata","poly_shp.shp",package = "ExtractTrainData"))
res_Col_name<-"Id"
ExValue_from_Polyshp(ras_img,poly_shp,res_Col_name)
}
\author{
Subhadip Datta
}
